Move parse to day8
